requirements.txt
streamlit
pytesseract
torch
transformers
Pillow
psycopg2-binary
PyYAML
streamlit-authenticator

